refactor(simulation): organize create flow by technology
