--- src/main/java/com/renewsim/backend/simulation_service/application/createSimulation/HydroSystemProfile.java
@@ -1,7 +1,0 @@
-package com.renewsim.backend.simulation_service.application.createSimulation;
-
-public record HydroSystemProfile(
-        double installedCapacityKw,
-        double degradationRateAnnualPct,
-        double availabilityPct) implements TechnologySystemProfile {
-}
