--- src/main/java/com/renewsim/backend/simulation_service/application/createSimulation/TechnologySystemProfile.java
@@ -1,4 +1,0 @@
-package com.renewsim.backend.simulation_service.application.createSimulation;
-
-public interface TechnologySystemProfile {
-}
