+++ src/main/java/com/renewsim/backend/simulation_service/create/application/technology/TechnologySystemProfile.java
@@ -1,0 +1,4 @@
+package com.renewsim.backend.simulation_service.create.application.technology;
+
+public interface TechnologySystemProfile {
+}
