+++ src/main/java/com/renewsim/backend/simulation_service/create/application/technology/solar/SolarSimulationEngine.java
@@ -1,10 +1,11 @@
-package com.renewsim.backend.simulation_service.application.createSimulation;
+package com.renewsim.backend.simulation_service.create.application.technology.solar;
 
 import com.renewsim.backend.simulation_service.application.port.out.PvgisSolarResourcePort;
 import com.renewsim.backend.simulation_service.application.shared.SimulationDetailsResult;
 import com.renewsim.backend.simulation_service.create.application.CreateRealSimulationCommand;
 import com.renewsim.backend.simulation_service.create.application.FinancialCalculator;
 import com.renewsim.backend.simulation_service.create.application.SimulationEngine;
+import com.renewsim.backend.simulation_service.create.application.technology.SimulationAssessmentPolicy;
 import com.renewsim.backend.simulation_service.domain.model.Simulation;
 import com.renewsim.backend.simulation_service.domain.model.vo.ConsumptionProfile;
 import com.renewsim.backend.simulation_service.domain.model.vo.Technology;
