+++ src/main/java/com/renewsim/backend/simulation_service/create/application/technology/wind/WindSimulationEngine.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.application.createSimulation;
+package com.renewsim.backend.simulation_service.create.application.technology.wind;
 
 import com.renewsim.backend.simulation_service.application.shared.SimulationDetailsResult;
 import com.renewsim.backend.simulation_service.create.application.CreateRealSimulationCommand;
@@ -9,20 +9,20 @@ import com.renewsim.backend.simulation_service.domain.model.vo.Technology;
 import org.springframework.stereotype.Component;
 
 @Component
-public class HydroSimulationEngine implements SimulationEngine {
+public class WindSimulationEngine implements SimulationEngine {
 
     @Override
     public boolean supports(Technology technology) {
-        return "hydro".equals(technology.value());
+        return "wind".equals(technology.value());
     }
 
     @Override
     public void assertImplemented() {
-        throw new InvalidSimulationTechnologyException("UNSUPPORTED_TECHNOLOGY: 'hydro' simulation is not implemented yet");
+        throw new InvalidSimulationTechnologyException("UNSUPPORTED_TECHNOLOGY: 'wind' simulation is not implemented yet");
     }
 
     @Override
     public SimulationDetailsResult simulate(Simulation simulation, CreateRealSimulationCommand command) {
-        throw new IllegalStateException("Hydro simulation should not execute before implementation is available.");
+        throw new IllegalStateException("Wind simulation should not execute before implementation is available.");
     }
 }
